+++ tests/Classifiers/MultilayerPerceptronTest.php
@@ -101,7 +101,7 @@ class MultilayerPerceptronTest extends TestCase
             new Noise(1e-5),
             new Dense(8),
             new Activation(new LeakyReLU(0.1)),
-        ], 32, new Adam(0.001), 1e-4, 100, 1e-3, 5, 0.1, new CrossEntropy(), new FBeta());
+        ], 32, new Adam(0.001), 100, 1e-3, 3, 5, 0.1, new CrossEntropy(), new FBeta());
 
         $this->metric = new FBeta();
 
@@ -175,9 +175,9 @@ class MultilayerPerceptronTest extends TestCase
             ],
             'batch size' => 32,
             'optimizer' => new Adam(0.001),
-            'l2 penalty' => 1e-4,
             'epochs' => 100,
             'min change' => 1e-3,
+            'eval interval' => 3,
             'window' => 5,
             'hold out' => 0.1,
             'cost fn' => new CrossEntropy(),